python/sample-app-oauth/requirements.txt
Flask>=2.0.0
python-dotenv>=0.19.0
gohighlevel-api-client==1.0.0b1
